Refactor useScene STORE
