+++ src/components/SceneContainer/TurnOfComponent.vue
@@ -12,7 +12,7 @@
 import { computed } from 'vue'
 import { useI18n } from 'vue-i18n'
 import { useGameStore } from '@/stores/useGame'
-import { useSceneStore } from '@/stores/useSceneStore'
+import { useSceneStore } from '@/stores/useScene'
 
 // STORE
 const sceneStore = useSceneStore()
