+++ src/components/director/DirectorStepsComponent.vue
@@ -46,7 +46,8 @@
 import { onMounted, computed, ref } from 'vue'
 import { useI18n } from 'vue-i18n'
 import { useCountdownStore } from '@/stores/useCountdownStore'
-import { Scene, useSceneStore } from '@/stores/useSceneStore'
+import { Scene } from '@/stores/useScene/types'
+import { useSceneStore } from '@/stores/useScene'
 import { EmittedEvent } from '@/events'
 import { DirectorStep } from '@/views/DirectorView/types/directorViewTypes'
 import AddNewSceneModal from '@/components/AddNewSceneModal.vue'
