+++ src/stores/useScene/types.ts
@@ -1,0 +1,15 @@
+export type Scene = {
+  id: number
+  title: string
+  instructions: string
+  duration: number
+  rndBtnOptions?: string[]
+  type: string
+  selected?: boolean
+  isOpenAccordion?: boolean
+}
+
+export interface SelectScenesObject {
+  numberOfScenes: number
+  shuffle: boolean
+}
